+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.3 - 2026-02-06
+
+### Added
+- Quick command to set rust-fmt as the default formatter, with Global or Workspace scope selection.
+- Smart prompt when Rust is not using rust-fmt as the default formatter.
+
+### Changed
+- Workspace formatting is now faster on large projects.
+
+-----
 ## 0.1.2 - 2026-01-28
 
 ### Added
@@ -20,4 +30,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Auto-detect `Cargo.toml` to set crate root and `--edition`.
 - Auto-detect `rustfmt.toml` / `.rustfmt.toml` and pass `--config-path`.
 - Auto-detect `rust-toolchain(.toml)` and set `RUSTUP_TOOLCHAIN` when running `rustfmt`.
-